Create Cohort Definition class object

This function creates a Cohort Definition class object from multiple component parts. A cohort definition contains at a minimum a primary criteria class. The cohort definition can further contain a inclusion rules, additional criteria, censoring criteria and end strategy classes to provide more details on cohort restriction and cohort exit. Other components may also be manipulated but since they do not rely on a concept set expressions, they can be manipulated in separate methods. The cohort definition class differs from the circe expression in that it does not have a separate space for concept set expressions, which are bundled within the component.

createCohortDefinition(
  Name,
  Description = NA_character_,
  Author = NA_character_,
  cdmVersionRange = ">=5.0.0",
  PrimaryCriteria,
  AdditionalCriteria = NULL,
  InclusionRules = NULL,
  EndStrategy = NULL,
  CensoringCriteria = NULL,
  CohortEra = NULL
)

Arguments

Name
make a name for the cohort to add to the cohort details

Description
add a description detail to cohort details, optional

Author
add an author name to cohort details, optional

cdmVersionRange
add a cdm version range typically >= 5.0.0, please specify if not v5

PrimaryCriteria
add primary criteria object

AdditionalCriteria
add additional criteria object. if null then will create an additional criteria with qualified limit

InclusionRules
add inclusion rules object. if null will create empty inclusion rules with expression limit

EndStrategy
add end strategy object. if null will add end of continuous era strategy

CensoringCriteria
add censoring criteria object. if null will add empty censoring criteria

CohortEra
add cohort era object. if null will add collapse settings with 0 day pad and no censor window

Value

cohort definition class object with defined inputs. This can now be compiled into ohdisql and converted to json