Function to edit Primary Criteria

This function edits a meta data class

editPrimaryCriteria(
  primaryCriteria,
  detail = c("Name", "Description", "CriteriaList", "Attribute", "PriorDays",
    "PostDays", "ObservationWindow", "PrimaryCriteriaLimit", "ConceptSetItem",
    "ConceptMapping"),
  edit,
  add = FALSE,
  index = NULL,
  mapping = NULL
)

Arguments

primaryCriteria
the primary criteria component you wish to edit

detail
the slot to edit. The options include: Name, Description, CriteriaList, Attribute, PriorDays, PostDays, ObservationWindow, PrimaryCriteriaLimit, ConceptSetItem, ConceptSetMapping. Each slot has a particular edit type.

edit
the edit to make. If the detail is Name or Description the edit must be a character string. If the edit is PriorDays, PostDays or ObservationWindow the edit must be an integer, where the ObservationWindow is an edit of two integers to modify both the prior and post days. If the edit is to the PrimaryCriteriaLimit the edit must be a character string of All, First or Last. If the edit is to the conceptSetItem the edit must be a ConceptSetItem class. And if the edit is to the concept set mapping the edit must be a logical (T/F). If the edit is to the CriteriaList it must be a query type component and if it is to the attribute it must be an attribute type component

add
a loggic toggle to say if you are adding a piece to the pc component

index
an index to specify the position in a vector. This is needed for CriteriaList, Attribute, and edits to the concept sets. The CriteriaList only needs a single index. The others need one index for the position in the list and a second for the position inside the substructure.

mapping
an character string specifying the mapping to change. Options are includeDescendants, isExcluded, and includeMapped. This is only required if the detail is ConceptSetMapping

Value

the edit s4 class object