Fill in Concept Set details using a vocab

Concept sets created in R using the `cs` function do not contain details like "CONCEPT_NAME", "DOMAIN_ID", etc. If an OMOP CDM vocabulary is available then these details can be filled in by the the `getConceptSetDetails` function.

getConceptSetDetails(x, con, vocabularyDatabaseSchema = NULL)

Arguments

x

A concept set created by `cs`

con

A connection to an OMOP CDM database

vocabularyDatabaseSchema

Schema name where your OMOP vocabulary format resides. Note that for SQL Server, this should include both the database and schema name, for example 'vocabulary.dbo'.

Value

A modified version of the input concept set with concept details filled in.

Examples

if (FALSE) {
# create a concept set
vocabularyDatabaseSchema = "cdm5"
anemia <- cs(descendants(439777,4013073,4013074))

# fill in the details from an OMOP CDM
library(DatabaseConnector)
con <- connect(dbms = "postgresql", user = "postgres", password = "", server = "localhost/covid")
anemia <- getConceptSetDetails(condition_anemia, con, vocabularyDatabaseSchema = "cdm5")
}