Lookup concept name as a general search

This function looks up concepts based on the concept name. It can be modified to conduct an exact name search or general search that contains the concept name in the concept.

lookupKeyword(
  keyword,
  searchType = c("like", "exact", "any"),
  connectionDetails = NULL,
  connection = NULL,
  vocabularyDatabaseSchema = NULL,
  tempEmulationSchema = NULL
)

Arguments

keyword

a character string used to search OMOP concepts

searchType

options to aid search. Can use like match, exact match or any match

connectionDetails

An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection

An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

vocabularyDatabaseSchema

Schema name where your OMOP vocabulary format resides. Note that for SQL Server, this should include both the database and schema name, for example 'vocabulary.dbo'.

tempEmulationSchema

Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

Value

a tibble data frame object with conceptId, conceptName, standardConcept, standardConceptCaption, invalidReason, invalidReasonCaption, conceptCode, domainId, vocabularyId, conceptClassId.